requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
SQLAlchemy==2.0.49
asyncpg==0.31.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
pytz==2025.2
bcrypt==5.0.0
typing_extensions==4.15.0
